requirements.txt
msal==1.28.0
requests==2.31.0
azure-storage-blob==12.19.0
python-dotenv==1.0.0
anthropic==0.25.0
